Seperate Projects Header from Project page
